クラスを別ファイルに
