+++ modules/__init__.py
@@ -1,0 +1,1 @@
+from .calendar_event import CalendarEvent
